bugs and testing

+++ hooks.c
@@ -13,7 +13,7 @@ struct myfs_hook_operations myfs_hook_ops = {
 int myfs_hook_reg_create(int (*fun)(struct inode * inode)){
 	myfs_hook_ops.create_inode = fun;
 	printk("myfs_hook_reg_create called.\n");
-	printk("%d\n",fun);
+	printk("%d\n",(int)fun);
 	myfs_hook_ops.create_inode(NULL);
 	fun(NULL);
 	return -ENOSPC;
